requirements.txt
fastapi
nba_api
pandas
python-dotenv
psycopg2-binary
uvicorn